+++ src/components/Carousel/CarouselItem.tsx
@@ -19,10 +19,11 @@ const CarouselItem = forwardRef<HTMLDivElement, CarouselItemProps>(({ className,
   return (
     <div
       ref={ref}
-      role="group"
+      role={props.onClick ? 'button' : 'group'}
       aria-roledescription="slide"
       className={classNames(`${baseClassName}-item`, className, {
         [`${baseClassName}-item--gap-${columnGap}`]: !!columnGap,
+        [`${baseClassName}-item--cursor-pointer`]: !!props.onClick,
       })}
       {...props}
       {...commonProps}