+++ src/components/Navigation/NavigationItemTrigger/NavigationItemTrigger.tsx
@@ -18,8 +18,9 @@ const MobileNavigationItemTrigger = ({ id, label, children }: NavigationItemTrig
   return (
     <Accordion>
       <AccordionItem
-        variation="sm"
-        id={id ?? `${label}-accordion`}
+        hasTransition
+        key={`accordion-key-${label}`}
+        id={`accordion-item-${id}`}
         label={<Text variant={TextVariants.snwHeaderLink}>{label}</Text>}
       >
         {children}