+++ src/components/TextSymbol/index.ts
@@ -1,0 +1,2 @@
+export { TextSymbolVariants } from './types';
+export { default as TextSymbols, type TextSymbolProps } from './TextSymbol';