+++ src/patterns/LanguageSelector/LanguageSelector.tsx
@@ -28,7 +28,7 @@ const MobileLanguageSelector = ({
   return (
     <Accordion {...props}>
       <AccordionItem
-        variation="sm"
+        hasTransition
         id={id ?? 'language-selector-accordion'}
         label={<Text variant={TextVariants.snwHeaderLink}>{label}</Text>}
       >